+++ events.c
@@ -1828,10 +1828,10 @@ void HandlePropertyNotify(void)
 				x = GetIconOffset(icon);
 				icon->bm_w =
 				        XCreateWindow(dpy, icon->w, x, 0,
-				                      (unsigned int) icon->width,
-				                      (unsigned int) icon->height,
-				                      (unsigned int) 0, Scr->d_depth,
-				                      (unsigned int) CopyFromParent, Scr->d_visual,
+				                      icon->width,
+				                      icon->height,
+				                      0, Scr->d_depth,
+				                      CopyFromParent, Scr->d_visual,
 				                      valuemask, &attributes);
 
 				if(!(Tmp_win->wmhints->flags & IconMaskHint)) {
